fix: GitLab URL verification regex is now valid
The regex used to ensure name paths for gitlab.com repositories
incorrectly ordered a capture class to put a hyphen ('-') between a
period ('.') and a plus ('+') which indicated a character range match
instead of looking for a match for any of the characters individually.
Moved the hyphen to be the last character to avoid this.

Formatted rust code according to clippy's suggestions.