feat: Email can be hashed to generate colour for profile icon
Profile icons are to be generated from a contributors primary email such
that each profile is unique but does not rely on an external service
(currently gravatar). The colour is generated from the hash of the email
while the initials of the username for a contributor is displayed over
the solid colour. May adapt to generate custom bitmap art from email
hash.

Swapped bitmap and bitmap hash fields from contributor struct to colour
and initials for this first implementation.