feat: Added timeouts for key validation and AI summaries
When internet disconnection occurs while AI summaries are being generated or API key is being validated, there is now a timeout that once runs out allows the user to know if the generation/validation fails.

Error messages are specific to the given errors. Not general and outline the specific error that occurs within a reasonable scope.

Fixed some of the syntax to also use match cases rather than just checking for the existance of a response.

AI summary errors now handled with a flag on front-end .svelte file rather than just checking for existence of the error message.

NOTE: For the AI summary generation fails, still need to bubble it up properly with error message. Currently just displays error message in contributer cards repeatedly. Might need to change expected output for errors on the function of summarize_commits() from reqwest::Error to String.