feat: Added error handling for incorrect .csv/.tsv format
Appropriate error displays when the format for the grading sheet is not followed.
Save button is disabled, when this happens.

Added a handful of try-catch blocks for when "compute_preview_from_bytes" function is used, as this is where the format is checked.
When the "compute_preview_from_bytes" detects error in format, the given component variables are changed and displayed.

Error message is simple for user to understand, and stands out on the modal when error occurs. Error messages display next to save button.

Error messages now display in debug terminal when failed upload occurs.

All other functionality is maintained, i.e. adding a new file with correct format works after failed upload occurs.